encoder/requirements.txt
sentence-transformers==2.2.2
psycopg2-binary==2.9.7
torch==2.0.1
transformers==4.33.2
numpy==1.24.3 